+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 * Optional inline SVG workaround ([#1](https://github.com/marp-team/marpit/pull/1))
+* Split the injection of markdown-it plugins and provide interface of markdown-it plugin ([#2](https://github.com/marp-team/marpit/pull/2))
 
 ## v0.0.0 - 2018-03-24